requirements.txt
fastapi
uvicorn
python-multipart
requests
pylint
flake8
astroid
python-dotenv
geminiapi
